refactor(init): extract enrichResponse into shared module and export it
Move the response enrichment logic (args→command, run_neon_init→run_shell_command)
into lib/enrich-output.ts and export it alongside routeDataStep so neonctl
can apply the same transforms when routing --data step requests.
